Decrease annotations jittering

+++ src/celengine/render.cpp
@@ -5816,7 +5816,7 @@ Renderer::renderAnnotationMarker(const Annotation &a,
 
     glColor(a.color);
     glPushMatrix();
-    glTranslatef((int)a.position.x(), (int)a.position.y(), depth);
+    glTranslatef(a.position.x(), a.position.y(), depth);
 
     if (markerRep.symbol() == MarkerRepresentation::Crosshair)
         renderCrosshair(size, realTime, a.color);
@@ -5843,8 +5843,8 @@ Renderer::renderAnnotationLabel(const Annotation &a,
 {
     glColor(a.color);
     glPushMatrix();
-    glTranslatef((int)a.position.x() + hOffset + PixelOffset,
-                 (int)a.position.y() + vOffset + PixelOffset,
+    glTranslatef(a.position.x() + hOffset + PixelOffset,
+                 a.position.y() + vOffset + PixelOffset,
                  depth);
     font[fs]->bind();
     font[fs]->render(a.labelText, 0.0f, 0.0f);
